convert wrapper into checker fucntion

+++ backend/app/api/project.py
@@ -2,7 +2,7 @@ from flask import Blueprint, current_app
 from flask_login import current_user, login_required
 from flask_restful import Api, Resource, abort, fields, marshal_with, reqparse
 
-from app.checker import check_create_project_param
+from app.checker import check_create_project_param, check_project_permission
 from app.service import ProjectService, UserService
 
 bp = Blueprint(
@@ -29,15 +29,19 @@ class Project(Resource):
         "project_language": fields.String,
         "creator_id": fields.Integer
     }
+    def __init__(self):
+        
+        super().__init__()
     
     @login_required
     @marshal_with(res_fields)
-    def get(self, proj_id):
-        res, flag = proj_service.get_project(proj_id)
-        if flag:
-            return res, 200
-        else:
-            abort(404, message="Project {} doesn't exist".format(proj_id))
+    def get(self, proj_id): 
+        if check_project_permission(proj_id, "read"):
+            res, flag = proj_service.get_project(proj_id)
+            if flag:
+                return res, 200
+            else:
+                abort(404, message="Project {} doesn't exist".format(proj_id))
 
     @login_required
     def post(self):
@@ -68,23 +72,24 @@ class Project(Resource):
 
     @login_required
     def delete(self, proj_id):
-        proj, flag = proj_service.get_project(proj_id)
-        if flag:
-            admins = proj.admin_users
-            flag = False
-            for admin in admins:
-                if admin.username == current_user.username:
-                    flag = True
+        if check_project_permission(proj_id, "admin"):
+            proj, flag = proj_service.get_project(proj_id)
             if flag:
-                res, flag = proj_service.remove_project(proj_id)
+                admins = proj.admin_users
+                flag = False
+                for admin in admins:
+                    if admin.username == current_user.username:
+                        flag = True
                 if flag:
-                    return '', 204
+                    res, flag = proj_service.remove_project(proj_id)
+                    if flag:
+                        return '', 204
+                    else:
+                        abort(500, message=res)
                 else:
-                    abort(500, message=res)
+                    abort(400, message="Permission denied")
             else:
-                abort(400, message="Permission denied")
-        else:
-            abort(404, message="Project {} doesn't exist".format(proj_id))
+                abort(404, message="Project {} doesn't exist".format(proj_id))
 
 
 api.add_resource(Project, '/api/project/<int:proj_id>', '/api/project/')
